chore: add version check code to not show warning on >=4.52
