+++ packages/qwik/src/optimizer/src/plugins/rollup.ts
@@ -17,6 +17,7 @@ import {
   type QwikPlugin,
   type QwikPluginOptions,
 } from './plugin';
+import { findDepPkgJsonPath } from './utils';
 
 type QwikRollupPluginApi = {
   getOptimizer: () => Optimizer;
@@ -75,7 +76,7 @@ export function qwikRollup(qwikRollupOpts: QwikRollupPluginOptions = {}): any {
     },
 
     outputOptions(rollupOutputOpts) {
-      return normalizeRollupOutputOptionsObject(qwikPlugin, rollupOutputOpts, false);
+      return normalizeRollupOutputOptionsObject(qwikPlugin, rollupOutputOpts, false) as any;
     },
 
     async buildStart() {
@@ -150,11 +151,11 @@ export function normalizeRollupOutputOptions(
   };
 }
 
-export function normalizeRollupOutputOptionsObject(
+export async function normalizeRollupOutputOptionsObject(
   qwikPlugin: QwikPlugin,
   rollupOutputOptsObj: Rollup.OutputOptions | undefined,
   useAssetsDir: boolean
-): Rollup.OutputOptions {
+): Promise<Rollup.OutputOptions> {
   const outputOpts: Rollup.OutputOptions = { ...rollupOutputOptsObj };
   const opts = qwikPlugin.getOptions();
   const optimizer = qwikPlugin.getOptimizer();
@@ -252,10 +253,32 @@ export function normalizeRollupOutputOptionsObject(
    */
   outputOpts.hoistTransitiveImports = false;
 
-  outputOpts.onlyExplicitManualChunks = true;
-  console.warn(
-    `⚠️ For the latest and greatest, we recommend to install rollup@^4.52.0 directly in your project. It enables the new Rollup \`outputOpts.onlyExplicitManualChunks\` feature flag, which improves preloading performance and reduces cache invalidation for a snappier user experience.`
-  );
+  const userPkgJsonPath = await findDepPkgJsonPath(optimizer.sys, 'rollup', optimizer.sys.cwd());
+  if (userPkgJsonPath) {
+    try {
+      const fs: typeof import('fs') = await optimizer.sys.dynamicImport('node:fs');
+      const pkgJsonStr = await fs.promises.readFile(userPkgJsonPath, 'utf-8');
+      const pkgJson = JSON.parse(pkgJsonStr);
+      const version = String(pkgJson?.version || '');
+      const [major, minor, patch] = version.split('.').map((n: string) => parseInt(n, 10)) as [
+        number,
+        number,
+        number,
+      ];
+      const isGte452 =
+        Number.isFinite(major) &&
+        (major > 4 || (major === 4 && (minor > 52 || (minor === 52 && (patch || 0) >= 0))));
+      if (isGte452) {
+        outputOpts.onlyExplicitManualChunks = true;
+      } else {
+        console.warn(
+          `⚠️ For the latest and greatest, we recommend to install rollup@^4.52.0 directly in your project. It enables the new Rollup \`outputOpts.onlyExplicitManualChunks\` feature flag, which improves preloading performance and reduces cache invalidation for a snappier user experience.`
+        );
+      }
+    } catch {
+      // If we cannot determine the installed Rollup version, avoid warning
+    }
+  }
 
   return outputOpts;
 }
